feat: add safe service, resolve pr comment
