+++ packages/filler/src/services/ContractInteractionService.ts
@@ -34,7 +34,7 @@ import { CacheService } from "./CacheService"
 export class ContractInteractionService {
 	private configService: ChainConfigService
 	private api: ApiPromise | null = null
-	private cacheService: CacheService
+	public cacheService: CacheService
 
 	constructor(
 		private clientManager: ChainClientManager,
@@ -590,186 +590,5 @@ export class ContractInteractionService {
 		const totalBalanceUsd = nativeTokenUsdValue + daiBalanceUsd + usdtBalanceUsd + usdcBalanceUsd
 
 		return { nativeTokenBalance, daiBalance, usdtBalance, usdcBalance, totalBalanceUsd }
-	}
-
-	async calculateSwapOperations(
-		order: Order,
-		destChain: string,
-	): Promise<{ calls: { to: HexString; data: HexString; value: bigint }[]; totalGasEstimate: bigint }> {
-		// Check cache first
-		const cachedOperations = this.cacheService.getSwapOperations(order.id!)
-		if (cachedOperations) {
-			console.log(`Using cached swap operations for order ${order.id}`)
-			return {
-				calls: cachedOperations.calls.map((call) => ({
-					to: call.to as HexString,
-					data: call.data as HexString,
-					value: BigInt(call.value),
-				})),
-				totalGasEstimate: cachedOperations.totalGasEstimate,
-			}
-		}
-
-		const calls: { to: HexString; data: HexString; value: bigint }[] = []
-		let totalGasEstimate = BigInt(0)
-		const fillerWalletAddress = privateKeyToAddress(this.privateKey)
-		const destClient = this.clientManager.getPublicClient(destChain)
-
-		const daiAsset = this.configService.getDaiAsset(destChain)
-		const usdtAsset = this.configService.getUsdtAsset(destChain)
-		const usdcAsset = this.configService.getUsdcAsset(destChain)
-		const daiDecimals = await this.getTokenDecimals(daiAsset, destChain)
-		const usdtDecimals = await this.getTokenDecimals(usdtAsset, destChain)
-		const usdcDecimals = await this.getTokenDecimals(usdcAsset, destChain)
-
-		for (const token of order.outputs) {
-			const tokenAddress = bytes32ToBytes20(token.token)
-			const { nativeTokenBalance, daiBalance, usdcBalance, usdtBalance } = await this.getFillerBalanceUSD(
-				order,
-				destChain,
-			)
-
-			// Get current balance of the required token
-			const currentBalance =
-				tokenAddress == daiAsset
-					? daiBalance
-					: tokenAddress == usdtAsset
-						? usdtBalance
-						: tokenAddress == usdcAsset
-							? usdcBalance
-							: nativeTokenBalance
-
-			// Calculate how much more we need (in actual uint256 with decimals)
-			const balanceNeeded = token.amount > currentBalance ? token.amount - currentBalance : BigInt(0)
-
-			if (balanceNeeded > BigInt(0)) {
-				// Convert all balances to the same unit (18 decimals) for comparison
-				const normalizedBalances = {
-					dai: daiBalance / BigInt(10 ** daiDecimals),
-					usdt: usdtBalance / BigInt(10 ** usdtDecimals),
-					usdc: usdcBalance / BigInt(10 ** usdcDecimals),
-					native: nativeTokenBalance / BigInt(10 ** 18),
-				}
-
-				// Sort balances in descending order
-				const sortedBalances = Object.entries(normalizedBalances).sort(([, a], [, b]) => Number(b - a))
-
-				// Try to fulfill the requirement using the highest balance first
-				let remainingNeeded = balanceNeeded
-				for (const [tokenType, normalizedBalance] of sortedBalances) {
-					if (remainingNeeded <= BigInt(0)) break
-
-					// Skip if this is the same token we're trying to get
-					if (
-						(tokenType === "dai" && tokenAddress === daiAsset) ||
-						(tokenType === "usdt" && tokenAddress === usdtAsset) ||
-						(tokenType === "usdc" && tokenAddress === usdcAsset) ||
-						(tokenType === "native" && tokenAddress === ADDRESS_ZERO)
-					) {
-						continue
-					}
-
-					// Get the actual balance with decimals
-					const actualBalance =
-						tokenType === "dai"
-							? daiBalance
-							: tokenType === "usdt"
-								? usdtBalance
-								: tokenType === "usdc"
-									? usdcBalance
-									: nativeTokenBalance
-
-					// Calculate how much we can swap from this token (in actual uint256 with decimals)
-					const swapAmount = actualBalance > remainingNeeded ? remainingNeeded : actualBalance
-
-					if (swapAmount > BigInt(0)) {
-						const tokenToSwap =
-							tokenType === "dai"
-								? daiAsset
-								: tokenType === "usdt"
-									? usdtAsset
-									: tokenType === "usdc"
-										? usdcAsset
-										: ADDRESS_ZERO
-
-						const amountsIn = await destClient.readContract({
-							address: this.configService.getUniswapRouterV2Address(destChain),
-							abi: UNISWAP_ROUTER_V2_ABI,
-							functionName: "getAmountsIn",
-							args: [swapAmount, [tokenToSwap, tokenAddress]],
-						})
-
-						const amountIn = amountsIn[0]
-
-						const approveData = encodeFunctionData({
-							abi: ERC20_ABI,
-							functionName: "approve",
-							args: [this.configService.getUniswapRouterV2Address(destChain), amountIn],
-						})
-
-						const approveCall = {
-							to: tokenToSwap,
-							data: approveData,
-							value: BigInt(0),
-						}
-
-						const swapData = encodeFunctionData({
-							abi: UNISWAP_ROUTER_V2_ABI,
-							functionName: "swapTokensForExactTokens",
-							args: [
-								swapAmount,
-								amountIn,
-								[tokenToSwap, tokenAddress],
-								fillerWalletAddress,
-								order.deadline,
-							],
-						})
-
-						const call = {
-							to: this.configService.getUniswapRouterV2Address(destChain),
-							data: swapData,
-							value: BigInt(0),
-						}
-
-						try {
-							const { results } = await destClient.simulateCalls({
-								account: fillerWalletAddress,
-								calls: [approveCall, call],
-							})
-
-							const operationGasEstimate = results.reduce(
-								(acc, result) => acc + result.gasUsed,
-								BigInt(0),
-							)
-
-							calls.push(approveCall, call)
-							totalGasEstimate += operationGasEstimate
-							remainingNeeded -= swapAmount
-						} catch (simulationError) {
-							console.error(`Swap simulation failed for ${tokenType}:`, simulationError)
-							continue
-						}
-					}
-				}
-
-				// If we still need more tokens after trying all balances
-				if (remainingNeeded > BigInt(0)) {
-					throw new Error(`Insufficient balance to fulfill token requirement for ${tokenAddress}`)
-				}
-			}
-		}
-
-		// Cache the results
-		this.cacheService.setSwapOperations(
-			order.id!,
-			calls.map((call) => ({
-				to: call.to,
-				data: call.data,
-				value: call.value.toString(),
-			})),
-			totalGasEstimate,
-		)
-
-		return { calls, totalGasEstimate }
 	}
 }
